+++ source/renderer/app/themes/utils/createTheme.js
@@ -843,13 +843,9 @@ export const createDaedalusComponentsTheme = (
       '--theme-receive-qr-code-foreground-color': '#000',
     },
     recoveryPhrase: {
-      '--theme-recovery-phrase-normal-background-color':
-        'rgba(32, 34, 37, 0.05)',
-      '--theme-recovery-phrase-normal-border-color': 'rgba(68, 91, 124, .07)',
-      '--theme-recovery-phrase-warning-background-color':
-        'rgba(68, 91, 124, 0.05)',
-      '--theme-recovery-phrase-attention-background-color':
-        'rgba(234, 76, 91, .1)',
+      '--theme-recovery-phrase-attention-background-color': `${
+        background.primary.light
+      }`,
     },
     reportIssue: {
       '--theme-report-issue-button-background-color': `${