+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog
 - Reorganized Storybook by domain ([PR 1537](https://github.com/input-output-hk/daedalus/pull/1537))
 - Reorganized Tests by domain ([PR 1540](https://github.com/input-output-hk/daedalus/pull/1540))
 
+### Fixes
+
+- Fixed flat button color  ([PR 1586](https://github.com/input-output-hk/daedalus/pull/1586))
+
 ## 0.15.0
 
 ### Features